+++ internal/aws/proxy/server_test.go
@@ -29,8 +29,8 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/testutil"
 	"github.com/stretchr/testify/assert"
-	"go.opentelemetry.io/collector/testutil"
 	"go.uber.org/zap"
 )
