+++ receiver/awsecscontainermetricsreceiver/internal/awsecscontainermetrics/utils.go
@@ -19,7 +19,7 @@ import (
 	"time"
 
 	metricspb "github.com/census-instrumentation/opencensus-proto/gen-go/metrics/v1"
-	"go.opentelemetry.io/collector/testutil/metricstestutil"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/metricstestutil"
 	"google.golang.org/protobuf/types/known/timestamppb"
 )
